Change NetworkInfo fields and constructors to use u64 for total_sent and total_received

+++ common-types/src/network.rs
@@ -1,11 +1,11 @@
-// todo-zm: change from u32 to u64
 #[derive(Debug, Clone)]
 pub struct NetworkInfo {
-    pub total_sent: u32,
-    pub total_received: u32,
+    pub total_sent: u64,
+    pub total_received: u64,
 }
+
 impl NetworkInfo {
-    pub fn new(total_sent: u32, total_received: u32) -> NetworkInfo {
+    pub fn new(total_sent: u64, total_received: u64) -> NetworkInfo {
         NetworkInfo {
             total_sent,
             total_received,
